[FIX] pia: recompute allowed_implantation_ids at implantation creation
As the allowed_implantation_ids computed field has not any depends
fields, this fields was not recomputed when a new implantation was
created. This lead to an access error if this implantation was used in
another record.

This commit force the re computation for the current user at each new
implantation creation in order  to avoid those unwanted access errors.